backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.0
supabase>=2.5.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
spacy>=3.7.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
